tplink-smarthome: Reject packets that don't match
If this dissector is going to be registered to a port that isn't
IANA assigned to it, it should at least reject packets that don't
look like the protocol. If it's going to handle TCP desegmentation,
it also should deal with not necessarily starting at the beginning
of a stream too.


(cherry picked from commit 223748d23c7f8be1c2c3ce362c72c088b1794455)